(refactor) move htmlx2jsx handlers into own files
