+++ packages/svelte2tsx/src/htmlxtojsx/nodes/await.ts
