+++ packages/svelte2tsx/src/nodes/slot.ts
@@ -10,7 +10,7 @@ import {
 } from '../utils/svelteAst';
 import TemplateScope from './TemplateScope';
 import { SvelteIdentifier, WithName, BaseDirective } from '../interfaces';
-import { getTypeForComponent } from '../htmlxtojsx/nodes/component-type';
+import { getTypeForComponent } from '../htmlxtojsx/utils/node-utils';
 
 function attributeStrValueAsJsExpression(attr: Node): string {
     if (attr.value.length == 0) return "''"; //wut?
